Add SVG and PNG badges with white background
Introduces SVG and PNG badges with white background and an optional white border.

The fully transparent badges can be hard to place on a colored background. These badges are based on a version used on https://www.digitalpublicgoods.net/registry (to the left of the DPG counter), which has a white background and white border.

The SVG versions are based on the Illustrator files.